refactor: split paid HTTP protocol crates
