+++ crates/bloom-paid-http/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "bloom-paid-http"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+description = "Shared paid HTTP request lifecycle, challenge normalization, and policy helpers"
+
+[dependencies]
+bloom-proto.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+toml.workspace = true
+url.workspace = true
+reqwest.workspace = true
+mpp.workspace = true
