+++ crates/bloom-paid-mpp/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "bloom-paid-mpp"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+description = "Tempo MPP adapter for Bloom paid HTTP requests"
+
+[dependencies]
+bloom-paid-http.workspace = true
+bloom-keystore.workspace = true
+bloom-proto.workspace = true
+bloom-tools.workspace = true
+serde_json.workspace = true
+async-trait.workspace = true
+reqwest.workspace = true
+mpp.workspace = true
