+++ crates/bloom-paid-x402/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "bloom-paid-x402"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+description = "x402 adapter for Bloom paid HTTP requests"
+
+[dependencies]
+bloom-paid-http.workspace = true
+bloom-keystore.workspace = true
+serde_json.workspace = true
+alloy.workspace = true
+base64.workspace = true
