+++ crates/bloom-vfs/Cargo.toml
@@ -15,6 +15,9 @@ bloom-objects.workspace = true
 bloom-petal-manifest.workspace = true
 bloom-chain-types.workspace = true
 bloom-keystore.workspace = true
+bloom-paid-http.workspace = true
+bloom-paid-mpp.workspace = true
+bloom-paid-x402.workspace = true
 bloom-chain.workspace = true
 bloom-rpc.workspace = true
 bloom-tx.workspace = true
