+++ crates/bloom-vfs/src/handlers/requests.rs
@@ -3,27 +3,25 @@
 //! This handler owns the `/requests` VFS tree. Reads only expose durable
 //! artefacts; payment/signing boundaries are writable control files.
 
-use std::borrow::Cow;
-use std::collections::BTreeMap;
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::time::{SystemTime, UNIX_EPOCH};
 
-use alloy::dyn_abi::Eip712Domain;
-use alloy::primitives::{Address, B256, U256, keccak256};
-use alloy::signers::SignerSync;
-use alloy::sol;
-use alloy::sol_types::SolStruct;
 use async_trait::async_trait;
-use base64::Engine;
-use base64::engine::general_purpose::STANDARD;
-use bloom_keystore::{Keystore, KeystoreError, WalletKind};
+use bloom_keystore::Keystore;
+use bloom_paid_http::{
+    NormalizedChallenge, ParsedRequest, PaymentRequirement, PolicyCheck, PolicyEvalInput,
+    evaluate_payment_policy, evaluate_session_policy, headers_to_string_map, json_number,
+    normalize_challenge, paid_http_intent_label, parse_money, parse_request,
+    select_payment_requirement, trim_money,
+};
+use bloom_paid_mpp::{PaymentBackend, RealMppBackend};
+use bloom_paid_x402::{KeystoreX402PaymentSigner, X402PaymentSigner, X402SignContext};
 use bloom_proto::Policy;
-use mpp::client::{PaymentProvider, TempoProvider, TempoSessionProvider};
 use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 use serde::{Deserialize, Serialize};
 use serde_json::json;
+use std::sync::Arc;
+use std::time::{SystemTime, UNIX_EPOCH};
 use url::Url;
 
 use crate::handler::{Entry, Handler, HandlerError};
@@ -36,232 +34,6 @@ pub struct RequestsHandler {
     default_wallet: Option<String>,
     client: reqwest::Client,
     x402_signer: Arc<dyn X402PaymentSigner>,
-}
-
-pub trait X402PaymentSigner: Send + Sync {
-    fn sign_x402_payment(
-        &self,
-        ctx: &X402SignContext<'_>,
-    ) -> Result<X402PaymentCredential, HandlerError>;
-}
-
-pub struct X402SignContext<'a> {
-    pub wallet: &'a str,
-    /// Stable id of the staged pending request. The nonce in
-    /// `TransferWithAuthorization` is derived from this so it is bound to the
-    /// request the user actually confirmed, not a fresh id generated at sign
-    /// time.
-    pub request_id: &'a str,
-    pub request: &'a ParsedRequest,
-    pub challenge: &'a NormalizedChallenge,
-    pub requirement: &'a PaymentRequirement,
-}
-
-pub struct X402PaymentCredential {
-    /// The secret-bearing value sent as the `X-PAYMENT` retry header. This is
-    /// intentionally never persisted to credential.json.
-    pub header_value: String,
-    /// Redacted/public metadata safe to expose in the VFS.
-    pub public_metadata: serde_json::Value,
-}
-
-sol! {
-    #[allow(missing_docs)]
-    struct TransferWithAuthorization {
-        address from;
-        address to;
-        uint256 value;
-        uint256 validAfter;
-        uint256 validBefore;
-        bytes32 nonce;
-    }
-}
-
-pub struct KeystoreX402PaymentSigner {
-    keystore: Keystore,
-}
-
-impl KeystoreX402PaymentSigner {
-    pub fn new(keystore: Keystore) -> Self {
-        Self { keystore }
-    }
-}
-
-impl X402PaymentSigner for KeystoreX402PaymentSigner {
-    fn sign_x402_payment(
-        &self,
-        ctx: &X402SignContext<'_>,
-    ) -> Result<X402PaymentCredential, HandlerError> {
-        let signer = self.keystore.signer(ctx.wallet).map_err(|e| {
-            HandlerError::backend(x402_keystore_signer_error(ctx.wallet, &self.keystore, e))
-        })?;
-        let info = self.keystore.info(ctx.wallet).map_err(|e| {
-            HandlerError::backend(format!("x402 signer wallet metadata unavailable: {e}"))
-        })?;
-        let scheme = ctx.requirement.scheme.as_deref().unwrap_or("exact");
-        if scheme != "exact" {
-            return Err(HandlerError::backend(format!(
-                "x402 keystore signer supports exact EVM requirements, got scheme '{scheme}'"
-            )));
-        }
-        let network = ctx
-            .requirement
-            .network
-            .as_deref()
-            .ok_or_else(|| HandlerError::backend("x402 requirement missing network"))?;
-        let chain_id = x402_evm_chain_id(network)?;
-        let asset = parse_x402_address(ctx.requirement.asset.as_deref(), "asset")?;
-        let pay_to = parse_x402_address(ctx.requirement.pay_to.as_deref(), "payTo")?;
-        let now = unix_seconds();
-        let valid_after = U256::from(now.saturating_sub(600));
-        let valid_before = U256::from(now + x402_max_timeout_seconds(ctx.requirement));
-        let nonce = x402_nonce(ctx, now);
-        let value = U256::from_str_radix(ctx.requirement.amount.as_deref().unwrap_or("0"), 10)
-            .map_err(|e| HandlerError::backend(format!("x402 amount is not uint256: {e}")))?;
-        let authorization = json!({
-            "from": info.address.to_string(),
-            "to": pay_to.to_string(),
-            "value": value.to_string(),
-            "validAfter": valid_after.to_string(),
-            "validBefore": valid_before.to_string(),
-            "nonce": nonce.to_string(),
-        });
-        let auth = TransferWithAuthorization {
-            from: info.address,
-            to: pay_to,
-            value,
-            validAfter: valid_after,
-            validBefore: valid_before,
-            nonce,
-        };
-        let domain = Eip712Domain {
-            name: x402_requirement_extra_str(ctx.requirement, "name").map(Cow::Owned),
-            version: x402_requirement_extra_str(ctx.requirement, "version").map(Cow::Owned),
-            chain_id: Some(U256::from(chain_id)),
-            verifying_contract: Some(asset),
-            ..Eip712Domain::default()
-        };
-        let digest = auth.eip712_signing_hash(&domain);
-        let signature = signer
-            .sign_hash_sync(&digest)
-            .map_err(|e| HandlerError::backend(format!("x402 keystore signing failed: {e}")))?
-            .to_string();
-        let header = json!({
-            "x402Version": 1,
-            "scheme": scheme,
-            "network": network,
-            "payload": {
-                "signature": signature,
-                "authorization": authorization,
-            },
-        });
-        let header_bytes = serde_json::to_vec(&header)
-            .map_err(|e| HandlerError::backend(format!("serialize x402 header: {e}")))?;
-        Ok(X402PaymentCredential {
-            header_value: STANDARD.encode(header_bytes),
-            public_metadata: json!({
-                "signer_backend": "bloom-keystore",
-                "wallet": ctx.wallet,
-                "wallet_kind": wallet_kind_label(info.kind),
-                "address": info.address.to_string(),
-                "scheme": scheme,
-                "network": network,
-                "asset": asset.to_string(),
-                "pay_to": pay_to.to_string(),
-                "resource": ctx.requirement.resource.as_deref().unwrap_or(ctx.request.url.as_str()),
-                "authorization": authorization,
-                "signature": "redacted",
-            }),
-        })
-    }
-}
-
-fn parse_x402_address(value: Option<&str>, field: &str) -> Result<Address, HandlerError> {
-    value
-        .ok_or_else(|| HandlerError::backend(format!("x402 requirement missing {field}")))?
-        .parse::<Address>()
-        .map_err(|e| {
-            HandlerError::backend(format!(
-                "x402 requirement {field} is not an EVM address: {e}"
-            ))
-        })
-}
-
-fn x402_evm_chain_id(network: &str) -> Result<u64, HandlerError> {
-    match network {
-        "abstract" => Ok(2741),
-        "abstract-testnet" => Ok(11124),
-        "base-sepolia" => Ok(84532),
-        "base" => Ok(8453),
-        "avalanche-fuji" => Ok(43113),
-        "avalanche" => Ok(43114),
-        "iotex" => Ok(4689),
-        "sei" => Ok(1329),
-        "sei-testnet" => Ok(1328),
-        "polygon" => Ok(137),
-        "polygon-amoy" => Ok(80002),
-        "peaq" => Ok(3338),
-        "story" => Ok(1514),
-        "educhain" => Ok(41923),
-        "skale-base-sepolia" => Ok(324705682),
-        other => Err(HandlerError::backend(format!(
-            "x402 keystore signer supports EVM networks only; unsupported network '{other}'"
-        ))),
-    }
-}
-
-fn x402_requirement_extra_str(req: &PaymentRequirement, key: &str) -> Option<String> {
-    req.raw
-        .get("extra")
-        .and_then(|v| v.get(key))
-        .and_then(|v| v.as_str())
-        .map(str::to_string)
-}
-
-fn x402_max_timeout_seconds(req: &PaymentRequirement) -> u64 {
-    req.raw
-        .get("maxTimeoutSeconds")
-        .and_then(|v| v.as_u64())
-        .unwrap_or(60)
-}
-
-fn unix_seconds() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs()
-}
-
-fn x402_nonce(ctx: &X402SignContext<'_>, now: u64) -> B256 {
-    keccak256(
-        format!(
-            "{}:{}:{}:{now}",
-            ctx.wallet, ctx.request.url, ctx.request_id
-        )
-        .as_bytes(),
-    )
-}
-
-fn wallet_kind_label(kind: WalletKind) -> &'static str {
-    match kind {
-        WalletKind::Local => "local",
-        WalletKind::Watch => "watch",
-        WalletKind::PasskeyGated => "passkey",
-    }
-}
-
-fn x402_keystore_signer_error(wallet: &str, keystore: &Keystore, err: KeystoreError) -> String {
-    match err {
-        KeystoreError::Locked(_) => match keystore.info(wallet).map(|i| i.kind) {
-            Ok(WalletKind::PasskeyGated) => format!(
-                "wallet '{wallet}' is locked; passkey wallets must be foreground-unlocked with unlock_passkey before confirming this paid request"
-            ),
-            _ => format!(
-                "wallet '{wallet}' is locked; unlock the wallet before confirming this paid request"
-            ),
-        },
-        other => format!("x402 keystore signer unavailable for wallet '{wallet}': {other}"),
-    }
 }
 
 impl RequestsHandler {
@@ -341,7 +113,7 @@ impl RequestsHandler {
         self.ensure_layout()?;
         let text = std::str::from_utf8(input)
             .map_err(|_| HandlerError::invalid("request input must be UTF-8"))?;
-        let request = parse_request(text)?;
+        let request = parse_request(text).map_err(HandlerError::invalid)?;
         let id = new_request_id();
         let wallet = match self.select_wallet(request.wallet.as_deref()) {
             Ok(wallet) => wallet,
@@ -691,13 +463,16 @@ impl RequestsHandler {
                 "payment policy warning requires override sentinel '{sentinel}'"
             )));
         }
-        let credential = self.x402_signer.sign_x402_payment(&X402SignContext {
-            wallet: &wallet,
-            request_id: id,
-            request: &request,
-            challenge: &challenge,
-            requirement: &requirement,
-        })?;
+        let credential = self
+            .x402_signer
+            .sign_x402_payment(&X402SignContext {
+                wallet: &wallet,
+                request_id: id,
+                request: &request,
+                challenge: &challenge,
+                requirement: &requirement,
+            })
+            .map_err(HandlerError::backend)?;
 
         let mut retry = self.client.request(
             request.method.parse().unwrap_or(reqwest::Method::GET),
@@ -958,894 +733,8 @@ impl Handler for RequestsHandler {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct ParsedRequest {
-    method: String,
-    url: Url,
-    wallet: Option<String>,
-    max_amount_usd: Option<f64>,
-    headers: BTreeMap<String, String>,
-    body: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct TomlRequest {
-    method: Option<String>,
-    url: String,
-    wallet: Option<String>,
-    max_amount_usd: Option<String>,
-    #[serde(default)]
-    headers: BTreeMap<String, String>,
-    body: Option<TomlBody>,
-}
-
-#[derive(Debug, Deserialize)]
-struct TomlBody {
-    inline: Option<String>,
-}
-
-fn parse_request(input: &str) -> Result<ParsedRequest, HandlerError> {
-    let trimmed = input.trim();
-    if trimmed.is_empty() {
-        return Err(HandlerError::invalid("empty request"));
-    }
-    if trimmed.contains("url")
-        && trimmed.contains('=')
-        && let Ok(t) = toml::from_str::<TomlRequest>(trimmed)
-    {
-        return Ok(ParsedRequest {
-            method: t
-                .method
-                .unwrap_or_else(|| "GET".into())
-                .to_ascii_uppercase(),
-            url: Url::parse(&t.url).map_err(|e| HandlerError::invalid(format!("url: {e}")))?,
-            wallet: t.wallet,
-            max_amount_usd: t.max_amount_usd.as_deref().and_then(|v| v.parse().ok()),
-            headers: t.headers,
-            body: t.body.and_then(|b| b.inline),
-        });
-    }
-    let mut lines = trimmed.lines();
-    let first = lines.next().unwrap().trim();
-    let mut parts = first.split_whitespace();
-    let method = parts
-        .next()
-        .ok_or_else(|| HandlerError::invalid("missing method"))?
-        .to_ascii_uppercase();
-    let url = parts
-        .next()
-        .ok_or_else(|| HandlerError::invalid("missing URL"))?;
-    let mut wallet = None;
-    let mut max_amount_usd = None;
-    for attr in parts {
-        if let Some(v) = attr.strip_prefix("wallet=") {
-            wallet = Some(v.trim_matches('"').to_string());
-        }
-        if let Some(v) = attr.strip_prefix("max_amount_usd=") {
-            max_amount_usd = v.trim_matches('"').parse().ok();
-        }
-    }
-    let mut headers = BTreeMap::new();
-    let mut body_lines = Vec::new();
-    let mut in_body = false;
-    for line in lines {
-        if line.trim().is_empty() {
-            in_body = true;
-            continue;
-        }
-        if !in_body {
-            if let Some((k, v)) = line.split_once(':') {
-                headers.insert(k.trim().to_ascii_lowercase(), v.trim().to_string());
-            } else {
-                in_body = true;
-                body_lines.push(line);
-            }
-        } else {
-            body_lines.push(line);
-        }
-    }
-    Ok(ParsedRequest {
-        method,
-        url: Url::parse(url).map_err(|e| HandlerError::invalid(format!("url: {e}")))?,
-        wallet,
-        max_amount_usd,
-        headers,
-        body: if body_lines.is_empty() {
-            None
-        } else {
-            Some(body_lines.join("\n"))
-        },
-    })
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NormalizedChallenge {
-    protocol: String,
-    intent: String,
-    merchant: String,
-    realm: Option<String>,
-    network: Option<String>,
-    asset: Option<String>,
-    amount: Option<String>,
-    amount_usd: Option<f64>,
-    charge_id: Option<String>,
-    session_id: Option<String>,
-    deposit_amount: Option<String>,
-    deposit_usd: Option<f64>,
-    chain_id: Option<u64>,
-    unit_type: Option<String>,
-    channel_id: Option<String>,
-    challenge_id: Option<String>,
-    request: Option<serde_json::Value>,
-    headers: BTreeMap<String, String>,
-    accepts: Vec<PaymentRequirement>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PaymentRequirement {
-    pub scheme: Option<String>,
-    pub network: Option<String>,
-    pub asset: Option<String>,
-    pub amount: Option<String>,
-    pub pay_to: Option<String>,
-    pub resource: Option<String>,
-    pub raw: serde_json::Value,
-}
-
-impl NormalizedChallenge {
-    fn payment_method(&self) -> serde_json::Value {
-        json!({
-            "protocol": self.protocol,
-            "intent": self.intent,
-            "network": self.network,
-            "asset": self.asset,
-            "merchant": self.merchant,
-            "charge_id": self.charge_id,
-            "session_id": self.session_id,
-            "channel_id": self.channel_id,
-            "deposit_amount": self.deposit_amount,
-            "deposit_usd": self.deposit_usd,
-            "chain_id": self.chain_id,
-            "unit_type": self.unit_type,
-        })
-    }
-}
-
-fn normalize_challenge(headers: &HeaderMap, body: &[u8], url: &Url) -> NormalizedChallenge {
-    let header_map = headers_to_string_map(headers);
-    let www_values = headers
-        .get_all(reqwest::header::WWW_AUTHENTICATE)
-        .iter()
-        .filter_map(|v| v.to_str().ok())
-        .collect::<Vec<_>>();
-    if let Some((challenge, value)) = mpp::parse_www_authenticate_all(www_values)
-        .into_iter()
-        .filter_map(Result::ok)
-        .find(|c| c.method.as_str() == "tempo")
-        .and_then(|challenge| {
-            challenge
-                .request
-                .decode::<serde_json::Value>()
-                .ok()
-                .map(|value| (challenge, value))
-        })
-    {
-        let method_details = value
-            .get("methodDetails")
-            .unwrap_or(&serde_json::Value::Null);
-        let amount = json_string(&value, &["amount"]);
-        let deposit_amount = json_string(&value, &["suggestedDeposit"]);
-        let channel_id = json_string(method_details, &["channelId"])
-            .or_else(|| json_string(&value, &["sessionId"]));
-        return NormalizedChallenge {
-            protocol: "mpp".into(),
-            intent: challenge.intent.as_str().to_string(),
-            merchant: challenge.realm.clone(),
-            realm: Some(challenge.realm),
-            network: Some("tempo".into()),
-            asset: json_string(&value, &["currency"]),
-            amount_usd: amount.as_deref().and_then(parse_money),
-            amount,
-            charge_id: json_string(&value, &["externalId"]),
-            session_id: channel_id.clone(),
-            deposit_usd: deposit_amount.as_deref().and_then(parse_money),
-            deposit_amount,
-            chain_id: method_details.get("chainId").and_then(|v| v.as_u64()),
-            unit_type: json_string(&value, &["unitType"]),
-            channel_id,
-            challenge_id: Some(challenge.id),
-            request: Some(value),
-            headers: header_map,
-            accepts: Vec::new(),
-        };
-    }
-    let www = header_map
-        .get("www-authenticate")
-        .cloned()
-        .unwrap_or_default();
-    let lower_www = www.to_ascii_lowercase();
-    let body_json: serde_json::Value = serde_json::from_slice(body).unwrap_or_else(|_| json!({}));
-    let lower_body_protocol = json_string(&body_json, &["protocol", "paymentProtocol", "scheme"])
-        .unwrap_or_default()
-        .to_ascii_lowercase();
-    let body_type = json_string(&body_json, &["type", "kind", "challengeType"])
-        .unwrap_or_default()
-        .to_ascii_lowercase();
-
-    let protocol = if header_map.keys().any(|k| k.starts_with("x-payment"))
-        || body_json.get("x402Version").is_some()
-        || body_json.get("accepts").is_some()
-    {
-        "x402"
-    } else if lower_www.contains("tempo")
-        || lower_www.contains("mpp")
-        || lower_www.contains("payment")
-        || lower_body_protocol.contains("tempo")
-        || lower_body_protocol.contains("mpp")
-        || body_json.get("charge").is_some()
-        || body_json.get("session").is_some()
-    {
-        "mpp"
-    } else {
-        "unknown"
-    };
-    let intent = if protocol == "x402" {
-        "one_time"
-    } else if lower_www.contains("session")
-        || body_json.get("session").is_some()
-        || body_type == "session"
-    {
-        "session"
-    } else {
-        "charge"
-    };
-    let accepts = parse_payment_requirements(&body_json);
-    let session = body_json.get("session").unwrap_or(&serde_json::Value::Null);
-    let charge = body_json.get("charge").unwrap_or(&serde_json::Value::Null);
-    let network = json_string(&body_json, &["network"])
-        .or_else(|| json_string(session, &["network"]))
-        .or_else(|| json_string(charge, &["network"]))
-        .or_else(|| {
-            body_json
-                .pointer("/accepts/0/network")
-                .and_then(|v| v.as_str())
-                .map(str::to_string)
-        });
-    let asset = json_string(&body_json, &["asset", "currency"])
-        .or_else(|| json_string(session, &["asset", "currency"]))
-        .or_else(|| json_string(charge, &["asset", "currency"]))
-        .or_else(|| {
-            body_json
-                .pointer("/accepts/0/asset")
-                .and_then(|v| v.as_str())
-                .map(str::to_string)
-        });
-    let amount = if intent == "session" {
-        json_string(session, &["voucherAmount", "amount", "price", "cost"])
-            .or_else(|| json_string(&body_json, &["voucherAmount", "amount", "price", "cost"]))
-    } else {
-        json_string(charge, &["amount", "price", "cost"])
-            .or_else(|| json_string(&body_json, &["amount", "price", "cost"]))
-            .or_else(|| {
-                body_json
-                    .pointer("/accepts/0/maxAmountRequired")
-                    .and_then(|v| v.as_str())
-                    .map(str::to_string)
-            })
-    };
-    let amount_usd = if intent == "session" {
-        json_f64(session, &["voucherAmountUsd", "amountUsd", "usd"])
-            .or_else(|| json_f64(&body_json, &["voucherAmountUsd", "amountUsd", "usd"]))
-            .or_else(|| amount.as_deref().and_then(parse_money))
-    } else {
-        json_f64(charge, &["amountUsd", "usd"])
-            .or_else(|| json_f64(&body_json, &["amountUsd", "usd"]))
-            .or_else(|| {
-                body_json
-                    .pointer("/accepts/0/amountUsd")
-                    .and_then(json_number)
-            })
-            .or_else(|| {
-                body_json
-                    .pointer("/accepts/0/amount_usd")
-                    .and_then(json_number)
-            })
-            .or_else(|| amount.as_deref().and_then(parse_money))
-    };
-    let deposit_amount = json_string(session, &["depositAmount", "deposit", "topUpAmount"])
-        .or_else(|| json_string(&body_json, &["depositAmount", "deposit", "topUpAmount"]));
-    let deposit_usd = json_f64(
-        session,
-        &["depositAmountUsd", "depositUsd", "topUpAmountUsd"],
-    )
-    .or_else(|| {
-        json_f64(
-            &body_json,
-            &["depositAmountUsd", "depositUsd", "topUpAmountUsd"],
-        )
-    })
-    .or_else(|| deposit_amount.as_deref().and_then(parse_money));
-    let merchant = json_string(charge, &["merchant", "merchantId"])
-        .or_else(|| json_string(session, &["merchant", "merchantId"]))
-        .or_else(|| json_string(&body_json, &["merchant", "merchantId"]))
-        .unwrap_or_else(|| url.host_str().unwrap_or("unknown").into());
-
-    NormalizedChallenge {
-        protocol: protocol.into(),
-        intent: intent.into(),
-        merchant,
-        realm: extract_realm(&www),
-        network,
-        asset,
-        amount,
-        amount_usd,
-        charge_id: json_string(charge, &["id", "chargeId"])
-            .or_else(|| json_string(&body_json, &["chargeId"])),
-        session_id: json_string(session, &["id", "sessionId"])
-            .or_else(|| json_string(&body_json, &["sessionId"])),
-        deposit_amount,
-        deposit_usd,
-        chain_id: None,
-        unit_type: None,
-        channel_id: None,
-        challenge_id: None,
-        request: None,
-        headers: header_map,
-        accepts,
-    }
-}
-
-fn parse_payment_requirements(body_json: &serde_json::Value) -> Vec<PaymentRequirement> {
-    let accepts = body_json
-        .get("accepts")
-        .and_then(|v| v.as_array())
-        .cloned()
-        .unwrap_or_else(|| vec![body_json.clone()]);
-    accepts
-        .into_iter()
-        .filter(|v| v.is_object())
-        .map(|v| PaymentRequirement {
-            scheme: v.get("scheme").and_then(|x| x.as_str()).map(str::to_string),
-            network: v
-                .get("network")
-                .and_then(|x| x.as_str())
-                .map(str::to_string),
-            asset: v.get("asset").and_then(|x| x.as_str()).map(str::to_string),
-            amount: v
-                .get("maxAmountRequired")
-                .or_else(|| v.get("amount"))
-                .and_then(|x| x.as_str())
-                .map(str::to_string),
-            pay_to: v.get("payTo").and_then(|x| x.as_str()).map(str::to_string),
-            resource: v
-                .get("resource")
-                .and_then(|x| x.as_str())
-                .map(str::to_string),
-            raw: v,
-        })
-        .collect()
-}
-
-fn json_string(v: &serde_json::Value, keys: &[&str]) -> Option<String> {
-    keys.iter().find_map(|k| {
-        v.get(*k).and_then(|v| {
-            v.as_str()
-                .map(str::to_string)
-                .or_else(|| v.as_f64().map(trim_money))
-        })
-    })
-}
-
-fn json_f64(v: &serde_json::Value, keys: &[&str]) -> Option<f64> {
-    keys.iter().find_map(|k| {
-        v.get(*k)
-            .and_then(|v| v.as_f64().or_else(|| v.as_str().and_then(parse_money)))
-    })
-}
-
-fn parse_money(raw: &str) -> Option<f64> {
-    raw.trim().trim_start_matches('$').parse().ok()
-}
-
-fn trim_money(v: f64) -> String {
-    let s = format!("{v:.6}");
-    s.trim_end_matches('0').trim_end_matches('.').to_string()
-}
-
-fn paid_http_intent_label(intent: &str) -> &str {
-    match intent {
-        "charge" => "one_time",
-        other => other,
-    }
-}
-
-fn extract_realm(s: &str) -> Option<String> {
-    s.split(',').find_map(|part| {
-        part.trim()
-            .strip_prefix("realm=")
-            .map(|v| v.trim_matches('"').to_string())
-    })
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct PolicyCheck {
-    rule: String,
-    result: String,
-    detail: String,
-}
-
-#[derive(Debug, Clone, Copy)]
-struct PolicyEvalInput<'a> {
-    host: &'a str,
-    asset: Option<&'a str>,
-    network: Option<&'a str>,
-    intent: &'a str,
-    amount_usd: Option<f64>,
-    request_max_amount_usd: Option<f64>,
-    spent_24h_usd: f64,
-}
-
-fn evaluate_payment_policy(policy: &Policy, input: PolicyEvalInput<'_>) -> Vec<PolicyCheck> {
-    let mut out = Vec::new();
-    let payments = &policy.payments;
-    let host = input.host.to_ascii_lowercase();
-    let asset = input.asset;
-    let network = input.network;
-    let amount_usd = input.amount_usd;
-    push_check(
-        &mut out,
-        "payments.enabled",
-        payments.enabled,
-        if payments.enabled {
-            "paid HTTP enabled".into()
-        } else {
-            "wallet policy has not enabled paid HTTP".into()
-        },
-        false,
-    );
-    push_check(
-        &mut out,
-        "payments.require_plan",
-        payments.require_plan,
-        if payments.require_plan {
-            "staged plan is required before confirmation".into()
-        } else {
-            "paid HTTP requires a staged plan before confirmation".into()
-        },
-        false,
-    );
-
-    if contains_ci(&payments.http.deny_hosts, &host) {
-        out.push(deny("payments.http.deny_hosts", format!("{host} denied")));
-    }
-    if !payments.http.allow_hosts.is_empty() && !contains_ci(&payments.http.allow_hosts, &host) {
-        out.push(deny(
-            "payments.http.allow_hosts",
-            format!("{host} not allowed"),
-        ));
-    } else if !payments.http.allow_hosts.is_empty() {
-        out.push(pass("payments.http.allow_hosts", format!("{host} allowed")));
-    }
-
-    if let Some(asset) = asset {
-        if contains_ci(&payments.assets.deny, asset) {
-            out.push(deny("payments.assets.deny", format!("{asset} denied")));
-        }
-        if !payments.assets.allow.is_empty() && !contains_ci(&payments.assets.allow, asset) {
-            out.push(deny(
-                "payments.assets.allow",
-                format!("{asset} not allowed"),
-            ));
-        } else if !payments.assets.allow.is_empty() {
-            out.push(pass("payments.assets.allow", format!("{asset} allowed")));
-        }
-    }
-    if let Some(network) = network {
-        if contains_ci(&payments.networks.deny, network) {
-            out.push(deny("payments.networks.deny", format!("{network} denied")));
-        }
-        if !payments.networks.allow.is_empty() && !contains_ci(&payments.networks.allow, network) {
-            out.push(deny(
-                "payments.networks.allow",
-                format!("{network} not allowed"),
-            ));
-        } else if !payments.networks.allow.is_empty() {
-            out.push(pass(
-                "payments.networks.allow",
-                format!("{network} allowed"),
-            ));
-        }
-    }
-
-    if let Some(usd) = amount_usd {
-        if let Some(cap) = min_cap([policy.caps.per_tx_usd, payments.http.per_request_usd]) {
-            if usd > cap {
-                out.push(deny(
-                    "payments.http.per_request_usd",
-                    format_usd_cmp(usd, ">", cap),
-                ));
-            } else {
-                out.push(pass(
-                    "payments.http.per_request_usd",
-                    format_usd_cmp(usd, "<=", cap),
-                ));
-            }
-        }
-        if let Some(cap) = input.request_max_amount_usd {
-            if usd > cap {
-                out.push(deny(
-                    "request.max_amount_usd",
-                    format_usd_cmp(usd, ">", cap),
-                ));
-            } else {
-                out.push(pass(
-                    "request.max_amount_usd",
-                    format_usd_cmp(usd, "<=", cap),
-                ));
-            }
-        }
-        if let Some(cap) = min_cap([policy.caps.per_day_usd, payments.http.per_day_usd]) {
-            let total = input.spent_24h_usd + usd;
-            if total > cap {
-                out.push(deny(
-                    "payments.http.per_day_usd",
-                    format!("{} spent+request > {} cap", usd_fmt(total), usd_fmt(cap)),
-                ));
-            } else {
-                out.push(pass(
-                    "payments.http.per_day_usd",
-                    format!("{} spent+request <= {} cap", usd_fmt(total), usd_fmt(cap)),
-                ));
-            }
-        }
-        if let Some(warn) = policy.caps.require_confirm_above_usd {
-            if usd > warn {
-                out.push(warn_check(
-                    "caps.require_confirm_above_usd",
-                    format_usd_cmp(usd, ">", warn),
-                ));
-            } else {
-                out.push(pass(
-                    "caps.require_confirm_above_usd",
-                    format_usd_cmp(usd, "<=", warn),
-                ));
-            }
-        }
-        if input.intent == "session" {
-            if !payments.sessions.enabled {
-                out.push(deny(
-                    "payments.sessions.enabled",
-                    "session payments are disabled",
-                ));
-            } else {
-                out.push(pass(
-                    "payments.sessions.enabled",
-                    "session payments enabled",
-                ));
-            }
-            if let Some(cap) = payments.sessions.max_deposit_usd {
-                if usd > cap {
-                    out.push(deny(
-                        "payments.sessions.max_deposit_usd",
-                        format_usd_cmp(usd, ">", cap),
-                    ));
-                } else {
-                    out.push(pass(
-                        "payments.sessions.max_deposit_usd",
-                        format_usd_cmp(usd, "<=", cap),
-                    ));
-                }
-            }
-            if let Some(cap) = payments.sessions.max_session_spend_usd {
-                if usd > cap {
-                    out.push(deny(
-                        "payments.sessions.max_session_spend_usd",
-                        format_usd_cmp(usd, ">", cap),
-                    ));
-                } else {
-                    out.push(pass(
-                        "payments.sessions.max_session_spend_usd",
-                        format_usd_cmp(usd, "<=", cap),
-                    ));
-                }
-            }
-        }
-    } else {
-        out.push(warn_check(
-            "payments.amount_usd",
-            "paid HTTP challenge did not expose a USD-denominated amount; review before confirming",
-        ));
-    }
-
-    out
-}
-
-fn push_check(out: &mut Vec<PolicyCheck>, rule: &str, pass_ok: bool, detail: String, warn: bool) {
-    out.push(PolicyCheck {
-        rule: rule.into(),
-        result: if pass_ok {
-            "pass"
-        } else if warn {
-            "warn"
-        } else {
-            "deny"
-        }
-        .into(),
-        detail,
-    });
-}
-
-fn pass(rule: &str, detail: impl Into<String>) -> PolicyCheck {
-    PolicyCheck {
-        rule: rule.into(),
-        result: "pass".into(),
-        detail: detail.into(),
-    }
-}
-fn warn_check(rule: &str, detail: impl Into<String>) -> PolicyCheck {
-    PolicyCheck {
-        rule: rule.into(),
-        result: "warn".into(),
-        detail: detail.into(),
-    }
-}
-fn deny(rule: &str, detail: impl Into<String>) -> PolicyCheck {
-    PolicyCheck {
-        rule: rule.into(),
-        result: "deny".into(),
-        detail: detail.into(),
-    }
-}
-fn contains_ci(set: &std::collections::BTreeSet<String>, needle: &str) -> bool {
-    set.iter().any(|v| v.eq_ignore_ascii_case(needle))
-}
-fn min_cap<const N: usize>(values: [Option<f64>; N]) -> Option<f64> {
-    values.into_iter().flatten().reduce(f64::min)
-}
-fn format_usd_cmp(a: f64, op: &str, b: f64) -> String {
-    format!("{} {op} {}", usd_fmt(a), usd_fmt(b))
-}
-fn usd_fmt(v: f64) -> String {
-    format!("${v:.6}")
-        .trim_end_matches('0')
-        .trim_end_matches('.')
-        .to_string()
-}
-fn json_number(v: &serde_json::Value) -> Option<f64> {
-    v.as_f64().or_else(|| v.as_str()?.parse().ok())
-}
-
-fn select_payment_requirement(
-    challenge: &NormalizedChallenge,
-    policy: &Policy,
-    host: &str,
-) -> Option<PaymentRequirement> {
-    let candidates = if challenge.accepts.is_empty() {
-        vec![PaymentRequirement {
-            scheme: None,
-            network: challenge.network.clone(),
-            asset: challenge.asset.clone(),
-            amount: challenge.amount.clone(),
-            pay_to: None,
-            resource: None,
-            raw: json!({}),
-        }]
-    } else {
-        challenge.accepts.clone()
-    };
-    candidates.into_iter().find(|req| {
-        !evaluate_payment_policy(
-            policy,
-            PolicyEvalInput {
-                host,
-                asset: req.asset.as_deref(),
-                network: req.network.as_deref(),
-                intent: &challenge.intent,
-                amount_usd: challenge.amount_usd,
-                request_max_amount_usd: None,
-                spent_24h_usd: 0.0,
-            },
-        )
-        .iter()
-        .any(|c| c.result == "deny")
-    })
-}
-
-fn evaluate_session_policy(
-    policy: &Policy,
-    challenge: &NormalizedChallenge,
-    already_spent_usd: f64,
-) -> Vec<PolicyCheck> {
-    let mut out = Vec::new();
-    if challenge.intent != "session" {
-        return out;
-    }
-    let sessions = &policy.payments.sessions;
-    if !sessions.enabled {
-        out.push(PolicyCheck {
-            rule: "payments.sessions.enabled".into(),
-            result: "deny".into(),
-            detail: "wallet policy has not enabled payment sessions".into(),
-        });
-    } else {
-        out.push(PolicyCheck {
-            rule: "payments.sessions.enabled".into(),
-            result: "pass".into(),
-            detail: "payment sessions enabled".into(),
-        });
-    }
-    if let (Some(deposit), Some(cap)) = (challenge.deposit_usd, sessions.max_deposit_usd) {
-        out.push(PolicyCheck {
-            rule: "payments.sessions.max_deposit_usd".into(),
-            result: if deposit > cap { "deny" } else { "pass" }.into(),
-            detail: format!(
-                "{} {} {}",
-                trim_money(deposit),
-                if deposit > cap { ">" } else { "<=" },
-                trim_money(cap)
-            ),
-        });
-    }
-    if let (Some(amount), Some(cap)) = (challenge.amount_usd, sessions.max_session_spend_usd) {
-        let projected = already_spent_usd + amount;
-        out.push(PolicyCheck {
-            rule: "payments.sessions.max_session_spend_usd".into(),
-            result: if projected > cap { "deny" } else { "pass" }.into(),
-            detail: format!(
-                "projected cumulative {} {} {}",
-                trim_money(projected),
-                if projected > cap { ">" } else { "<=" },
-                trim_money(cap)
-            ),
-        });
-    }
-    out
-}
-
-#[async_trait]
-trait PaymentBackend: Send + Sync {
-    fn name(&self) -> &'static str;
-    async fn confirm(
-        &self,
-        challenge: &NormalizedChallenge,
-        request: &ParsedRequest,
-        wallet: &str,
-        policy: &Policy,
-        request_id: &str,
-    ) -> Result<PaymentExecution, HandlerError>;
-}
-
-struct RealMppBackend {
-    keystore: Keystore,
-    client: reqwest::Client,
-    rpc_url: String,
-}
-
-impl RealMppBackend {
-    fn signer_error(&self, wallet: &str, err: KeystoreError) -> HandlerError {
-        match err {
-            KeystoreError::Locked(_) => {
-                let kind = self
-                    .keystore
-                    .raw_policy(wallet)
-                    .ok()
-                    .map(|(_, kind)| kind)
-                    .or_else(|| self.keystore.info(wallet).ok().map(|info| info.kind));
-                if kind == Some(WalletKind::PasskeyGated) {
-                    HandlerError::invalid(format!(
-                        "passkey wallet '{wallet}' is locked; run the foreground passkey unlock flow (`unlock-passkey` / Keystore::unlock_passkey) before confirming Tempo MPP payments"
-                    ))
-                } else {
-                    HandlerError::invalid(format!(
-                        "wallet '{wallet}' is locked; unlock it before confirming Tempo MPP payments"
-                    ))
-                }
-            }
-            other => HandlerError::invalid(format!(
-                "wallet '{wallet}' cannot be used for Tempo MPP signing: {other}"
-            )),
-        }
-    }
-}
-
-struct PaymentExecution {
-    credential_metadata: serde_json::Value,
-    receipt_raw: serde_json::Value,
-    response_status: u16,
-    response_headers: HeaderMap,
-    response_body: Vec<u8>,
-}
-
 struct ConfirmResult {
     final_state: String,
-}
-
-#[async_trait]
-impl PaymentBackend for RealMppBackend {
-    fn name(&self) -> &'static str {
-        "mpp_tempo"
-    }
-
-    async fn confirm(
-        &self,
-        challenge: &NormalizedChallenge,
-        request: &ParsedRequest,
-        wallet: &str,
-        policy: &Policy,
-        _request_id: &str,
-    ) -> Result<PaymentExecution, HandlerError> {
-        if challenge.protocol != "mpp" || challenge.network.as_deref() != Some("tempo") {
-            return Err(HandlerError::invalid(
-                "only Tempo MPP challenges can be confirmed by the real MPP backend",
-            ));
-        }
-        let signer = self
-            .keystore
-            .signer(wallet)
-            .map_err(|e| self.signer_error(wallet, e))?;
-        let payment_challenge = parse_stored_mpp_challenge(challenge)?;
-        let credential = match challenge.intent.as_str() {
-            "charge" => {
-                let provider = TempoProvider::new((*signer).clone(), &self.rpc_url)
-                    .map_err(|e| HandlerError::backend(format!("TempoProvider: {e}")))?;
-                provider.pay(&payment_challenge).await
-            }
-            "session" => {
-                let mut provider = TempoSessionProvider::new((*signer).clone(), &self.rpc_url)
-                    .map_err(|e| HandlerError::backend(format!("TempoSessionProvider: {e}")))?;
-                if let Some(max) = policy
-                    .payments
-                    .sessions
-                    .max_deposit_usd
-                    .and_then(f64_to_u128_amount)
-                {
-                    provider = provider.with_max_deposit(max);
-                }
-                provider.pay(&payment_challenge).await
-            }
-            other => {
-                return Err(HandlerError::invalid(format!(
-                    "unsupported MPP intent '{other}'"
-                )));
-            }
-        }
-        .map_err(|e| HandlerError::backend(format!("Tempo MPP credential: {e}")))?;
-        let authorization = mpp::format_authorization(&credential)
-            .map_err(|e| HandlerError::backend(format!("format MPP Authorization: {e}")))?;
-        let authorization_sha256 = bloom_tools::sha256_hex(authorization.as_bytes());
-        let credential_value = serde_json::to_value(&credential).map_err(|e| {
-            HandlerError::backend(format!("serialize MPP credential metadata: {e}"))
-        })?;
-        let retry = retry_paid_request(&self.client, request, &authorization).await?;
-        let receipt_raw = retry
-            .headers
-            .get("payment-receipt")
-            .and_then(|h| h.to_str().ok())
-            .and_then(|h| mpp::parse_receipt(h).ok())
-            .and_then(|r| serde_json::to_value(r).ok())
-            .unwrap_or_else(|| json!({}));
-        Ok(PaymentExecution {
-            credential_metadata: json!({
-                "redacted": true,
-                "protocol": challenge.protocol,
-                "intent": challenge.intent,
-                "backend": self.name(),
-                "authorization_sha256": authorization_sha256,
-                "source": credential_value.get("source").cloned(),
-                "payload_type": credential_value.get("payload").and_then(|p| p.get("type")).cloned(),
-                "charge_id": challenge.charge_id,
-                "session_id": challenge.session_id,
-                "channel_id": challenge.channel_id,
-                "secret_material_in_vfs": false,
-                "raw_authorization_stored": false,
-                "raw_signed_payload_stored": false
-            }),
-            receipt_raw,
-            response_status: retry.status,
-            response_headers: retry.headers,
-            response_body: retry.body,
-        })
-    }
 }
 
 async fn confirm_with_backend(
@@ -1887,7 +776,8 @@ async fn confirm_with_backend(
     let policy = backend_policy_for_wallet(root, &wallet).unwrap_or_default();
     let execution = backend
         .confirm(&challenge, &request, &wallet, &policy, id)
-        .await?;
+        .await
+        .map_err(HandlerError::backend)?;
     let succeeded = execution.response_status < 400;
     let final_state = if succeeded { "sent" } else { "failed" };
     fs::create_dir_all(pending.join("response"))?;
@@ -1959,33 +849,6 @@ async fn confirm_with_backend(
     })
 }
 
-fn parse_stored_mpp_challenge(
-    challenge: &NormalizedChallenge,
-) -> Result<mpp::PaymentChallenge, HandlerError> {
-    challenge
-        .headers
-        .get("www-authenticate")
-        .and_then(|h| {
-            mpp::parse_www_authenticate_all([h.as_str()])
-                .into_iter()
-                .filter_map(Result::ok)
-                .find(|c| c.method.as_str() == "tempo" && c.intent.as_str() == challenge.intent)
-        })
-        .ok_or_else(|| {
-            HandlerError::backend(
-                "stored challenge is missing a parseable Tempo MPP WWW-Authenticate header",
-            )
-        })
-}
-
-fn f64_to_u128_amount(v: f64) -> Option<u128> {
-    if v.is_finite() && v >= 0.0 {
-        Some(v.floor() as u128)
-    } else {
-        None
-    }
-}
-
 fn parsed_request_from_artifact(v: &serde_json::Value) -> Result<ParsedRequest, HandlerError> {
     let method = v
         .get("method")
@@ -2013,50 +876,6 @@ fn parsed_request_from_artifact(v: &serde_json::Value) -> Result<ParsedRequest, 
 fn backend_policy_for_wallet(root: &Path, wallet: &str) -> Result<Policy, HandlerError> {
     let raw = fs::read_to_string(root.join("keystore").join(wallet).join("policy.toml"))?;
     toml::from_str(&raw).map_err(|e| HandlerError::backend(e.to_string()))
-}
-
-struct RetryResponse {
-    status: u16,
-    headers: HeaderMap,
-    body: Vec<u8>,
-}
-
-async fn retry_paid_request(
-    client: &reqwest::Client,
-    request: &ParsedRequest,
-    authorization: &str,
-) -> Result<RetryResponse, HandlerError> {
-    let mut req = client.request(
-        request.method.parse().unwrap_or(reqwest::Method::GET),
-        request.url.clone(),
-    );
-    for (k, v) in &request.headers {
-        let name = HeaderName::from_bytes(k.as_bytes())
-            .map_err(|e| HandlerError::invalid(format!("invalid header {k}: {e}")))?;
-        let val = HeaderValue::from_str(v)
-            .map_err(|e| HandlerError::invalid(format!("invalid header {k}: {e}")))?;
-        req = req.header(name, val);
-    }
-    req = req.header(reqwest::header::AUTHORIZATION, authorization);
-    if let Some(body) = &request.body {
-        req = req.body(body.clone());
-    }
-    let response = req
-        .send()
-        .await
-        .map_err(|e| HandlerError::backend(format!("paid HTTP retry failed: {e}")))?;
-    let status = response.status().as_u16();
-    let headers = response.headers().clone();
-    let body = response
-        .bytes()
-        .await
-        .map_err(|e| HandlerError::backend(format!("read paid HTTP retry response: {e}")))?
-        .to_vec();
-    Ok(RetryResponse {
-        status,
-        headers,
-        body,
-    })
 }
 
 fn request_json_field(dir: &Path, field: &str) -> serde_json::Value {
@@ -2206,17 +1025,6 @@ fn write_request_artifacts(
     Ok(())
 }
 
-fn headers_to_string_map(headers: &HeaderMap) -> BTreeMap<String, String> {
-    headers
-        .iter()
-        .map(|(k, v)| {
-            (
-                k.as_str().to_ascii_lowercase(),
-                v.to_str().unwrap_or("").to_string(),
-            )
-        })
-        .collect()
-}
 fn headers_to_json(headers: &HeaderMap) -> serde_json::Value {
     json!(headers_to_string_map(headers))
 }
@@ -2273,6 +1081,9 @@ fn new_request_id() -> String {
 mod tests {
     use super::*;
     use crate::path::VfsPath;
+    use bloom_paid_mpp::PaymentExecution;
+    use mpp::client::{PaymentProvider, TempoProvider};
+    use std::collections::BTreeMap;
     use std::sync::{
         Arc,
         atomic::{AtomicUsize, Ordering},
@@ -2748,7 +1559,7 @@ inline = '{"prompt":"hi"}'
             _wallet: &str,
             _policy: &Policy,
             _request_id: &str,
-        ) -> Result<PaymentExecution, HandlerError> {
+        ) -> Result<PaymentExecution, String> {
             Ok(PaymentExecution {
                 credential_metadata: json!({
                     "redacted": true,
@@ -2828,7 +1639,7 @@ inline = '{"prompt":"hi"}'
             _wallet: &str,
             _policy: &Policy,
             _request_id: &str,
-        ) -> Result<PaymentExecution, HandlerError> {
+        ) -> Result<PaymentExecution, String> {
             Ok(PaymentExecution {
                 credential_metadata: json!({
                     "redacted": true,
